+++ src/Altinn.App.Core/Internal/AccessManagement/Builders/DelegationBuilder.cs
@@ -1,9 +1,10 @@
 using System.Diagnostics.CodeAnalysis;
 using Altinn.App.Core.Helpers;
+using Altinn.App.Core.Internal.AccessManagement.Builders;
 using Altinn.App.Core.Internal.AccessManagement.Models;
 using Altinn.App.Core.Internal.AccessManagement.Models.Shared;
 
-namespace Altinn.App.Core.Internal.AccessManagement;
+namespace Altinn.App.Core.Internal.AccessManagement.Builders;
 
 internal abstract class DelegationBuilderBase
 {
